Added getControlIsDirty for Summary tab in app-entity-manage.
In app-entity-manage > summary tab > Things I need to fix widget has one of the business condition which has impacted attributes + impacted relationships then it will open 2 stepper business function dialog.

After that if user click on 'cancel' button which is skip in a way then it will go to second step for resolving relationship. But if user update any impacted attribute and click on 'save' button (Note: business function is 2 stepper so in this case save means next) it will open second stepper content where it shows relationships but in a fraction of seconds it close the dialog.

It should stay on second stepper. But it's getting closed and greyed overlay is not getting removed.

Because of this user has to reload application.

Problem:

Recently we have changed notification behavior where on govern complete notification previously it was showing toast message with reload button and now it's reloading 'rock-tab' automatically.